+++ mocks/import_documents.yaml
@@ -4,7 +4,7 @@ when:
   header:
     - name: X-TYPESENSE-API-KEY
       value: VerySecretKey
-  body: "{\"company_name\": \"test\", \"num_employees\": 1, \"country\": \"c1\"}\n{\"company_name\": \"test2\", \"num_employees\": 2, \"country\": \"c2\"}"
+  body: "{\"company_name\":\"test\",\"num_employees\":1,\"country\":\"c1\"}\n{\"company_name\":\"test2\",\"num_employees\":2,\"country\":\"c2\"}"
 then:
   status: 200
   header:
