+++ typesense/tests/api/documents.rs
@@ -1,9 +1,9 @@
 #![allow(dead_code)]
 
+use super::Config;
 use serde::{Deserialize, Serialize};
 use typesense::document::Document;
 use typesense::Typesense;
-use typesense_codegen::apis::configuration::{ApiKey, Configuration};
 use typesense_codegen::apis::documents_api;
 
 #[derive(Typesense, Serialize, Deserialize)]
@@ -15,7 +15,7 @@ struct Company {
     country: String,
 }
 
-async fn import_documents(host: impl Into<String>, api_key: impl Into<String>) {
+async fn import_documents() {
     let documents = [
         Company {
             company_name: "test".to_owned(),
@@ -31,17 +31,8 @@ async fn import_documents(host: impl Into<String>, api_key: impl Into<String>) {
     .map(|c| serde_json::to_string(&c).unwrap())
     .join("\n");
 
-    let config = Configuration {
-        base_path: host.into(),
-        api_key: Some(ApiKey {
-            prefix: None,
-            key: api_key.into(),
-        }),
-        ..Default::default()
-    };
-
     let resp = documents_api::import_documents(
-        &config,
+        Config::get(),
         &Company::collection_schema().name,
         documents,
         None,
@@ -58,12 +49,7 @@ mod tokio_test {
 
     #[tokio::test]
     async fn import_documents_tokio() {
-        let _ = dotenvy::dotenv();
-
-        let host = std::env::var("URL").expect("URL must be present in .env");
-        let api_key = std::env::var("API_KEY").expect("API_KEY must be present in .env");
-
-        import_documents(host, api_key).await
+        import_documents().await
     }
 }
 
@@ -74,13 +60,10 @@ mod wasm_test {
 
     wasm_bindgen_test::wasm_bindgen_test_configure!(run_in_browser);
 
-    const HOST: &str = "http://localhost:5000";
-    const API_KEY: &str = "VerySecretKey";
-
     #[wasm_bindgen_test]
     async fn import_documents_wasm() {
         console_error_panic_hook::set_once();
 
-        import_documents(HOST, API_KEY).await
+        import_documents().await
     }
 }
